Move InstrumentUtils more centrally to LoggerUtils
